8286925: Move JSON parser used in JFR tests to test library
