+++ test/jdk/jdk/jfr/tool/TestPrintJSON.java
@@ -35,7 +35,8 @@ import jdk.jfr.ValueDescriptor;
 import jdk.jfr.consumer.RecordedEvent;
 import jdk.jfr.consumer.RecordedObject;
 import jdk.jfr.consumer.RecordingFile;
-import jdk.jfr.tool.JSONValue.JSONArray;
+import jdk.test.lib.json.JSONValue;
+import jdk.test.lib.json.JSONValue.JSONArray;
 import jdk.test.lib.Asserts;
 import jdk.test.lib.process.OutputAnalyzer;
